Revert "Fix the native boot issue on GP2.0."
This reverts commit 40e4c8deca6784a2a200d80b00ec8596ef37dba0.

The original patch (enable RAID) actually didn't fix the native boot issue. It is another
from acrn GVT fixed the boot:

====================
commit 6ecee7901a93b712fe3ef425281c151f565366d4
Author: Fei Jiang <fei.jiang@intel.com>
Date: Tue Jun 12 11:15:21 2018 +0800

drm/i915/gvt: fix i915 boot issue when gvt is not enabled

And the enabling of RAID actually costs 20ms more for boot time which is not necessary.

=====================

So this patch suggest revert the original patch. The same request sent to PKT (through PKT-1009).